repository updated

+++ TaskPlannerApp/src/main/java/com/masai/repository/SprintRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.masai.model.Sprint;
 
 @Repository
-public interface SprintRepository extends JpaRepository<Sprint, Integer>{
+public interface SprintRepository extends JpaRepository<Sprint, Long>{
 
 }
